Boost performance by actually using the prepared calibration curve.
In case of multiple radiocarbon dates supplied, the desired behaviour
is to prepare the calibration curve once, before going through the
list of radiocarbon dates, and pass the same prepared object for
calibrating each date.

Unfortunately, before this commit the `calibrate()` polymorphic function
was called with `options.curve` as argument, and repeated the creation
of the calibration curve for each date.

This change seems to provide a free performance improvement of ~6-7%
on the calibration procedure (i.e. excluding plotting).